Restore internal ledge linkage and inline addon deallocation
