+++ README.md
@@ -1,6 +1,6 @@
 # _saga_
 
-![Progress](https://img.shields.io/badge/matching-28.87%25-red)
+![Progress](https://img.shields.io/badge/matching-28.88%25-red)
 [![Bazel build](https://github.com/ttdecomp/saga/actions/workflows/build-bazel.yaml/badge.svg)](https://github.com/ttdecomp/saga/actions/workflows/build-bazel.yaml)
 [![Discord](https://img.shields.io/discord/1467775700894224555?color=%235865F2&logo=discord&logoColor=%23FFFFFF)](https://discord.gg/2HJuMtzA7q)
 [![status & wasm build](https://img.shields.io/badge/status%20%26%20wasm%20build-click%20here-orange?style=flat)](https://opensaga.dev/)
@@ -73,11 +73,11 @@ See https://ttdecomp.github.io/saga/
 | `legoapi/core` | 25.9% | 6.6% |
 | `legoapi/cutscenes` | 29.9% | 3.9% |
 | `legoapi/gizmo` | 31.4% | 7.3% |
-| `legoapi/gizmos` | 47.4% | 32.8% |
+| `legoapi/gizmos` | 46.9% | 32.7% |
 | `legoapi/items` | 25.7% | 5.8% |
 | `legoapi/menus` | 17.7% | 6.2% |
 | `legoapi/misc` | 20.8% | 4.1% |
-| `legoapi/props` | 36.2% | 2.2% |
+| `legoapi/props` | 37.1% | 2.3% |
 | `legoapi/render` | 24.5% | 7.5% |
 | `legoapi/world` | 25.1% | 6.4% |
 | `legogame` | 61.9% | 8.8% |
